Update Constants.cs
Pulled each class out into their own individual files in constants/
