+++ game/scripts/constants/Actions.cs
@@ -1,0 +1,11 @@
+using Godot;
+
+namespace Chess;
+
+public static class Actions
+{
+	public static readonly StringName DeselectPiece = "DeselectPiece";
+	public static readonly StringName Interact = "Interact";
+	public static readonly StringName RotateCamera = "RotateCamera";
+	public static readonly StringName ToggleMenu = "ToggleMenu";
+}
