+++ game/scripts/constants/ChessMaterials.cs
@@ -1,0 +1,8 @@
+namespace Chess;
+
+public static class ChessMaterials
+{
+	public const string Black = $"{ResourcePaths.Chess}material-piece-black-4k.tres";
+	public const string Board = $"{ResourcePaths.Chess}material-board-4k.tres";
+	public const string White = $"{ResourcePaths.Chess}material-piece-white-4k.tres";
+}
